chore: move `helper.ts` from `cost-analysis/lib/` to `cost-explorer/lib`
