+++ src/services/cost-explorer/cost-analysis/modules/CostAnalysisChart.vue
@@ -125,9 +125,6 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { DEFAULT_CHART_COLORS, DISABLED_LEGEND_COLOR } from '@/styles/colorsets';
 
-import {
-    convertFilterItemToQueryStoreFilter,
-} from '@/services/cost-explorer/cost-analysis/lib/helper';
 import CostAnalysisPieChart
     from '@/services/cost-explorer/cost-analysis/modules/CostAnalysisPieChart.vue';
 import CostAnalysisStackedColumnChart
@@ -135,6 +132,9 @@ import CostAnalysisStackedColumnChart
 import {
     FILTER_ITEM_MAP, GRANULARITY,
 } from '@/services/cost-explorer/lib/config';
+import {
+    convertFilterItemToQueryStoreFilter,
+} from '@/services/cost-explorer/lib/helper';
 import CostExplorerFilterTags from '@/services/cost-explorer/modules/CostExplorerFilterTags.vue';
 import { costExplorerStore } from '@/services/cost-explorer/store';
 import type {
