+++ src/services/cost-explorer/cost-analysis/modules/CostAnalysisDataTable.vue
@@ -93,11 +93,11 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { ASSET_INVENTORY_ROUTE } from '@/services/asset-inventory/route-config';
 import {
-    convertFilterItemToQueryStoreFilter, getDataTableCostFields, getTimeUnitByPeriod,
-} from '@/services/cost-explorer/cost-analysis/lib/helper';
-import {
     FILTER, GRANULARITY, GROUP_BY, GROUP_BY_ITEM_MAP,
 } from '@/services/cost-explorer/lib/config';
+import {
+    convertFilterItemToQueryStoreFilter, getDataTableCostFields, getTimeUnitByPeriod,
+} from '@/services/cost-explorer/lib/helper';
 import { costExplorerStore } from '@/services/cost-explorer/store';
 import type { GroupByItem } from '@/services/cost-explorer/store/cost-analysis/type';
 import type { CostQueryFilterItem } from '@/services/cost-explorer/type';
