+++ src/services/cost-explorer/cost-analysis/modules/CostAnalysisQueryFilter.vue
@@ -46,8 +46,6 @@
 </template>
 
 <script lang="ts">
-
-
 import { computed, reactive, toRefs } from 'vue';
 
 import {
@@ -56,13 +54,12 @@ import {
 import type { MenuItem } from '@spaceone/design-system/dist/src/inputs/context-menu/type';
 import dayjs from 'dayjs';
 
-
 import { i18n } from '@/translations';
 
-import { getInitialDates } from '@/services/cost-explorer/cost-analysis/lib/helper';
 import CostAnalysisPeriodSelectDropdown
     from '@/services/cost-explorer/cost-analysis/modules/CostAnalysisPeriodSelectDropdown.vue';
 import { GRANULARITY } from '@/services/cost-explorer/lib/config';
+import { getInitialDates } from '@/services/cost-explorer/lib/helper';
 import CurrencySelectDropdown from '@/services/cost-explorer/modules/CurrencySelectDropdown.vue';
 import { costExplorerStore } from '@/services/cost-explorer/store';
 import type { Period } from '@/services/cost-explorer/type';
