+++ src/services/cost-explorer/cost-analysis/modules/CostAnalysisStackedColumnChart.vue
@@ -8,7 +8,6 @@
 </template>
 
 <script lang="ts">
-
 import {
     getCurrentInstance,
     reactive, toRefs, watch,
@@ -26,9 +25,7 @@ import { cloneDeep } from 'lodash';
 
 import { commaFormatter, numberFormatter } from '@cloudforet/core-lib';
 
-
 import { CURRENCY } from '@/store/modules/display/config';
-
 
 import config from '@/lib/config';
 import { currencyMoneyFormatter } from '@/lib/helper/currency-helper';
@@ -37,8 +34,8 @@ import { useProxyValue } from '@/common/composables/proxy-state';
 
 import { gray } from '@/styles/colors';
 
-import { getTimeUnitByPeriod } from '@/services/cost-explorer/cost-analysis/lib/helper';
 import { GRANULARITY } from '@/services/cost-explorer/lib/config';
+import { getTimeUnitByPeriod } from '@/services/cost-explorer/lib/helper';
 import type { Granularity } from '@/services/cost-explorer/type';
 import {
     getStackedChartData, getCurrencyAppliedChartData,
