+++ src/services/cost-explorer/store/cost-analysis/actions.ts
@@ -6,8 +6,8 @@ import { store } from '@/store';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { getInitialDates } from '@/services/cost-explorer/cost-analysis/lib/helper';
 import { GRANULARITY } from '@/services/cost-explorer/lib/config';
+import { getInitialDates } from '@/services/cost-explorer/lib/helper';
 import type { CostAnalysisStoreState } from '@/services/cost-explorer/store/cost-analysis/type';
 import type {
     CostQuerySetModel, CostQuerySetOption, CostQueryFilterItem, CostQueryFilters,
