+++ src/services/cost-explorer/widgets/modules/CostDashboardStackedColumnWidget.vue
@@ -59,10 +59,10 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { gray } from '@/styles/colors';
 
+import { GRANULARITY, GROUP_BY_ITEM_MAP } from '@/services/cost-explorer/lib/config';
 import {
     convertFilterItemToQueryStoreFilter,
-} from '@/services/cost-explorer/cost-analysis/lib/helper';
-import { GRANULARITY, GROUP_BY_ITEM_MAP } from '@/services/cost-explorer/lib/config';
+} from '@/services/cost-explorer/lib/helper';
 import type { CostQueryFilterItem, GroupBy, Period } from '@/services/cost-explorer/type';
 import {
     getCurrencyAppliedChartData, getLegends, getTooltipText, getXYChartData,
